@hauntsaninja 2023-02-26
Thanks again for the PR!

I got some time to benchmark the current version of this PR. On your lorem ipsum, it was about a 16% improvement from max. On internal datasets, it was somewhere between 5-20% better.

I tried benchmarking the version with PCRE, but I couldn't get it to work. I ran into `undefined symbol: _pcre2_script_run_8` / `symbol not found in flat namespace (__pcre2_script_run_8)` on a work machine / laptop; is it obvious to you what the issue is there?

Re the macro, the following compiles and seems to perform equivalently (the trick is rustc wants more curly braces to let you put the attribute on the closure), but I didn't confirm the assembly looks the same:

<details>

```diff
diff --git a/tiktoken/src/lib.rs b/tiktoken/src/lib.rs
index b44d9c8b21..40241be910 100644
--- a/tiktoken/src/lib.rs
+++ b/tiktoken/src/lib.rs
@@ -26,10 +26,9 @@ fn _byte_pair_merge<T>(
     // A closure also cannot capture a reference to `piece` without
     // the borrow checker complaining about the mutable borrows during
     // the assignments later in this code.
-    macro_rules! get_rank {
-        ($start_idx:expr, $skip:expr) => {{
-            let start_idx: usize = $start_idx;
-            let skip: usize = $skip;
+    let get_rank = {
+        #[inline(always)]
+        |parts: &Vec<(usize, usize)>, start_idx: usize, skip: usize| {
             if (start_idx + skip + 2) < parts.len() {
                 ranks
                     .get(&piece[parts[start_idx].0..parts[start_idx + skip + 2].0])
@@ -37,16 +36,13 @@ fn _byte_pair_merge<T>(
             } else {
                 None
             }
-        }};
-        ($idx:expr) => {{
-            get_rank!($idx, 0)
-        }};
-    }
+        }
+    };
 
     // We look up the ranks once in the beggining and iteratively update
     // them during each merge, which reduces the number of rank lookups.
     for i in 0..parts.len() - 2 {
-        match get_rank!(i) {
+        match get_rank(&parts, i, 0) {
             Some(rank) => {
                 // usize::MAX is a sentinel value and cannot be a valid rank
                 debug_assert!(rank != usize::MAX);
@@ -89,9 +85,9 @@ fn _byte_pair_merge<T>(
             // parts[i] and parts[i-1] before removing, which could thrash
             // the cache. Thus, we update the rank calculation by skipping over
             // parts[i + 1], by invoking `get_rank!` with `skip = 1`.
-            parts[i].1 = get_rank!(i, 1).unwrap_or(usize::MAX);
+            parts[i].1 = get_rank(&parts, i, 1).unwrap_or(usize::MAX);
             if i > 0 {
-                parts[i - 1].1 = get_rank!(i - 1, 1).unwrap_or(usize::MAX);
+                parts[i - 1].1 = get_rank(&parts, i - 1, 1).unwrap_or(usize::MAX);
             }
 
             parts.remove(i + 1);
```
</details>